Adjust ps1 scripts add Avalonia demo
