+++ GNU.Gettext/Examples.HelloAvalonia/po-compile-translations.ps1
