+++ GNU.Gettext/Examples.HelloAvalonia/po/fr.po
@@ -1,15 +1,11 @@
-# Messages français pour GNU gettext.
-# Copyright © 2006 Yoyodyne, Inc.
-# Michel Robitaille <robitail@IRO.UMontreal.CA>, 2006.
-# Christophe Combelles <ccomb@free.fr>, 2006
-# Serguei Tarassov <st@arbinada.com>, 2012
+# This file is distributed under the same license as the package.
 #
 msgid ""
 msgstr ""
 "Project-Id-Version: GNU gettext .NET\n"
 "Report-Msgid-Bugs-To: bug-gnu-gettext@gnu.org\n"
-"POT-Creation-Date: 2012-06-28 14:56:59+0200\n"
-"PO-Revision-Date: 2012-07-05 10:15+0100\n"
+"POT-Creation-Date: 2012-06-28 14:26:54+0200\n"
+"PO-Revision-Date: 2012-07-03 12:37+0100\n"
 "Last-Translator: Serguei TARASSOV <st@arbinada.com>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -18,24 +14,15 @@ msgstr ""
 "Language: fr\n"
 "Plural-Forms: nplurals=2; plural=(n > 1);\n"
 
-#: ../Form1.cs:87
-#: ../Form1.cs:120
+#: ../Program.cs:33
 msgid "Hello, world!"
 msgstr "Bonjour, le monde !"
 
-#: ../Form1.cs:92
-msgid "en-US"
-msgstr ""
+#: ../Program.cs:48
+msgid "Text encoding"
+msgstr "Codification du texte"
 
-#: ../Form1.cs:99
-msgid "fr-FR"
-msgstr ""
-
-#: ../Form1.cs:106
-msgid "ru-RU"
-msgstr ""
-
-#: ../Form1.cs:174
+#: ../Program.cs:50
 msgid ""
 "Here is an example of how one might continue a very long string\n"
 "for the common case the string represents multi-line output.\n"
@@ -43,18 +30,14 @@ msgstr ""
 "Voici une exemple d'une chaîne longue\n"
 "en cas commun elle est représentée par plusieurs lignes.\n"
 
-#: ../Form1.cs:42
-msgid "Text encoding"
-msgstr "Codification du texte"
-
-#: ../Form1.cs:29
+#: ../Program.cs:34
 #, csharp-format
 msgid "This program is running as process number \"{0}\"."
 msgstr "Ce programme est exécuté en tant que processus numéro \"{0}\"."
 
-#: ../Form1.cs:32
-#: ../Form1.cs:35
-#: ../Form1.cs:38
+#: ../Program.cs:37
+#: ../Program.cs:40
+#: ../Program.cs:43
 #, csharp-format
 msgid "found {0} similar word"
 msgid_plural "found {0} similar words"
@@ -62,26 +45,14 @@ msgstr[0] "{0} mot similaire trouvé"
 msgstr[1] "{0} mots similaires trouvés"
 
 #. Context: Computers
-#: ../Form1.cs:40
+#: ../Program.cs:46
 msgctxt "Computers"
 msgid "Text encoding"
 msgstr "Encodage du texte"
 
 #. Context: Military
-#: ../Form1.cs:41
+#: ../Program.cs:47
 msgctxt "Military"
 msgid "Text encoding"
 msgstr "Chiffrement du texte"
-
-#: ../Form1.cs:113
-msgid "Switch to English"
-msgstr "Basculer en Anglais"
-
-#: ../Form1.cs:114
-msgid "Switch to French"
-msgstr "Basculer en Français"
-
-#: ../Form1.cs:115
-msgid "Switch to Russian"
-msgstr "Basculer en Russe"
