+++ GNU.Gettext/Examples.HelloAvalonia/po/ru.po
@@ -1,12 +1,11 @@
-# Copyright (C) 2004 Free Software Foundation, Inc.
-# This file is distributed under the same license as the hello-csharp package.
+# This file is distributed under the same license as the package.
 #
 msgid ""
 msgstr ""
 "Project-Id-Version: Examples.Hello\n"
 "Report-Msgid-Bugs-To: bug-gnu-gettext@gnu.org\n"
-"POT-Creation-Date: 2012-06-28 14:56:59+0200\n"
-"PO-Revision-Date: 2012-07-05 10:16+0100\n"
+"POT-Creation-Date: 2012-06-28 14:26:54+0200\n"
+"PO-Revision-Date: 2012-06-28 14:29+0100\n"
 "Last-Translator: Serguei TARASSOV <st@arbinada.com>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -18,24 +17,15 @@ msgstr ""
 "X-Poedit-Country: RUSSIAN FEDERATION\n"
 "X-Poedit-SourceCharset: utf-8\n"
 
-#: ../Form1.cs:87
-#: ../Form1.cs:120
+#: Program.cs:32
 msgid "Hello, world!"
 msgstr "Здравствуй, мир!"
 
-#: ../Form1.cs:92
-msgid "en-US"
-msgstr ""
+#: Program.cs:47
+msgid "Text encoding"
+msgstr "Кодирование текста"
 
-#: ../Form1.cs:99
-msgid "fr-FR"
-msgstr ""
-
-#: ../Form1.cs:106
-msgid "ru-RU"
-msgstr ""
-
-#: ../Form1.cs:174
+#: Program.cs:49
 msgid ""
 "Here is an example of how one might continue a very long string\n"
 "for the common case the string represents multi-line output.\n"
@@ -43,19 +33,13 @@ msgstr ""
 "Вот пример как можно продолжить длинную фразу\n"
 "в общем случае она располагается на нескольких строках.\n"
 
-#: ../Form1.cs:42
-msgid "Text encoding"
-msgstr "Кодирование текста"
-
-#: ../Form1.cs:29
-#, csharp-format
+#: Program.cs:33
 msgid "This program is running as process number \"{0}\"."
 msgstr "Эта программа запущена под процессом номер \"{0}\"."
 
-#: ../Form1.cs:32
-#: ../Form1.cs:35
-#: ../Form1.cs:38
-#, csharp-format
+#: Program.cs:36
+#: Program.cs:39
+#: Program.cs:42
 msgid "found {0} similar word"
 msgid_plural "found {0} similar words"
 msgstr[0] "{0} похожее слово найдено"
@@ -63,26 +47,14 @@ msgstr[1] "{0} похожих слова найдено"
 msgstr[2] "{0} похожих слов найдено"
 
 #. Context: Computers
-#: ../Form1.cs:40
+#: Program.cs:45
 msgctxt "Computers"
 msgid "Text encoding"
 msgstr "Кодировка текста"
 
 #. Context: Military
-#: ../Form1.cs:41
+#: Program.cs:46
 msgctxt "Military"
 msgid "Text encoding"
 msgstr "Шифрование текста"
-
-#: ../Form1.cs:113
-msgid "Switch to English"
-msgstr "Переключиться на английский"
-
-#: ../Form1.cs:114
-msgid "Switch to French"
-msgstr "Переключиться на французский"
-
-#: ../Form1.cs:115
-msgid "Switch to Russian"
-msgstr "Переключиться на русский"
