--- GNU.Gettext/Examples.HelloForms/Examples.HelloForms.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <ProjectGuid>{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}</ProjectGuid>
-    <OutputType>WinExe</OutputType>
-    <RootNamespace>GNU.Gettext.Examples</RootNamespace>
-    <AssemblyName>Examples.HelloForms</AssemblyName>
-    <TargetFramework>net6.0-windows7.0</TargetFramework>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Include="po\fr.po" />
-    <None Include="po\ru.po" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\GNU.Gettext\GNU.Gettext.csproj">
-      <Name>GNU.Gettext</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\GNU.Gettext.Avalonia\GNU.Gettext.Avalonia.csproj">
-      <Name>GNU.Gettext.Avalonia</Name>
-    </ProjectReference>
-  </ItemGroup>
-</Project>
