--- GNU.Gettext/Examples.HelloForms/po-compile-translations.sh
@@ -1,5 +1,0 @@
-#!bin/sh
-
-# Compile PO files to satellite assemblies
-mono ./../../Bin/Debug/GNU.Gettext.Msgfmt.exe -l fr-FR -d ./bin/Debug -r Examples.HelloForms -L ./../../Bin/Debug ./po/fr.po
-mono ./../../Bin/Debug/GNU.Gettext.Msgfmt.exe -l ru-RU -d ./bin/Debug -r Examples.HelloForms -L ./../../Bin/Debug ./po/ru.po
