--- GNU.Gettext/Examples.HelloForms/po-make-template.sh
@@ -1,3 +1,0 @@
-#!bin/sh
-## Extract messages to PO template file
-mono ./../../Bin/Debug/GNU.Gettext.Xgettext.exe -D "./" --recursive -o "./po/Messages.pot"
