+++ GNU.Gettext/GNU.Gettext.Test/GNU.Gettext.Test.csproj
@@ -21,10 +21,6 @@
       <Project>{69B39E65-093E-417A-BFC9-B1CF7B981680}</Project>
       <Name>GNU.Gettext.Xgettext</Name>
     </ProjectReference>
-    <ProjectReference Include="..\GNU.Gettext.Avalonia\GNU.Gettext.Avalonia.csproj">
-      <Project>{A458C869-B3C4-4130-8D3F-00F44548DD5C}</Project>
-      <Name>GNU.Gettext.Avalonia</Name>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <Content Include="Data\Test01.po">
