--- GNU.Gettext/GNU.Gettext.WinForms/GNU.Gettext.WinForms.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <ProjectGuid>{A458C869-B3C4-4130-8D3F-00F44548DD5C}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <TargetFramework>net6.0-windows7.0</TargetFramework>
-    <UseWindowsForms>true</UseWindowsForms>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\GNU.Gettext\GNU.Gettext.csproj">
-      <Project>{6996D99C-2E0F-4EDA-9FF0-BB45F81EADC8}</Project>
-      <Name>GNU.Gettext</Name>
-    </ProjectReference>
-  </ItemGroup>
-</Project>
