+++ GNU.Gettext/GNU.Gettext.sln
@@ -1,23 +1,21 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33717.318
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GNU.Gettext", "GNU.Gettext\GNU.Gettext.csproj", "{6996D99C-2E0F-4EDA-9FF0-BB45F81EADC8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GNU.Gettext", "GNU.Gettext\GNU.Gettext.csproj", "{6996D99C-2E0F-4EDA-9FF0-BB45F81EADC8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GNU.Gettext.Msgfmt", "GNU.Gettext.Msgfmt\GNU.Gettext.Msgfmt.csproj", "{2E98861E-D84B-4F9D-9000-EC9744770769}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GNU.Gettext.Msgfmt", "GNU.Gettext.Msgfmt\GNU.Gettext.Msgfmt.csproj", "{2E98861E-D84B-4F9D-9000-EC9744770769}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GNU.Getopt", "..\GNU.Getopt\GNU.Getopt.csproj", "{24BE81B3-C22E-448B-A892-AAF95AB28747}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GNU.Getopt", "..\GNU.Getopt\GNU.Getopt.csproj", "{24BE81B3-C22E-448B-A892-AAF95AB28747}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Examples.Hello", "Examples.Hello\Examples.Hello.csproj", "{5B5BB7C9-4972-4269-92AC-E041F3A2DF0F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Examples.Hello", "Examples.Hello\Examples.Hello.csproj", "{5B5BB7C9-4972-4269-92AC-E041F3A2DF0F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Examples.HelloForms", "Examples.HelloForms\Examples.HelloForms.csproj", "{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GNU.Gettext.Test", "GNU.Gettext.Test\GNU.Gettext.Test.csproj", "{40A18DF7-3926-4AC4-8726-581F5C39082D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GNU.Gettext.Test", "GNU.Gettext.Test\GNU.Gettext.Test.csproj", "{40A18DF7-3926-4AC4-8726-581F5C39082D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GNU.Gettext.Xgettext", "GNU.Gettext.Xgettext\GNU.Gettext.Xgettext.csproj", "{69B39E65-093E-417A-BFC9-B1CF7B981680}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GNU.Gettext.Xgettext", "GNU.Gettext.Xgettext\GNU.Gettext.Xgettext.csproj", "{69B39E65-093E-417A-BFC9-B1CF7B981680}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GNU.Gettext.Avalonia", "GNU.Gettext.Avalonia\GNU.Gettext.Avalonia.csproj", "{A458C869-B3C4-4130-8D3F-00F44548DD5C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Examples.HelloAvalonia", "Examples.HelloAvalonia\Examples.HelloAvalonia.csproj", "{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -75,18 +73,6 @@ Global
 		{5B5BB7C9-4972-4269-92AC-E041F3A2DF0F}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{5B5BB7C9-4972-4269-92AC-E041F3A2DF0F}.Release|x86.ActiveCfg = Release|Any CPU
 		{5B5BB7C9-4972-4269-92AC-E041F3A2DF0F}.Release|x86.Build.0 = Release|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|x86.Build.0 = Debug|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|x86.ActiveCfg = Release|Any CPU
-		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|x86.Build.0 = Release|Any CPU
 		{40A18DF7-3926-4AC4-8726-581F5C39082D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{40A18DF7-3926-4AC4-8726-581F5C39082D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{40A18DF7-3926-4AC4-8726-581F5C39082D}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -111,18 +97,18 @@ Global
 		{69B39E65-093E-417A-BFC9-B1CF7B981680}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{69B39E65-093E-417A-BFC9-B1CF7B981680}.Release|x86.ActiveCfg = Release|Any CPU
 		{69B39E65-093E-417A-BFC9-B1CF7B981680}.Release|x86.Build.0 = Release|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Debug|x86.Build.0 = Debug|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Release|x86.ActiveCfg = Release|Any CPU
-		{A458C869-B3C4-4130-8D3F-00F44548DD5C}.Release|x86.Build.0 = Release|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Debug|x86.Build.0 = Debug|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|x86.ActiveCfg = Release|Any CPU
+		{D7DC509C-6BD5-4627-92A8-5CF39A2A41BC}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
